Guardado en base de datos/Mostrar todos los libros

+++ literalura/src/main/java/com/challenge/literalura/models/DatosLibro.java
@@ -10,6 +10,6 @@ public record DatosLibro(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutor> autor,
         @JsonAlias("languages") List<String> idioma,
-        @JsonAlias("download_count") Integer numeroDeDescargas
+        @JsonAlias("download_count") double numeroDeDescargas
 ) {
 }
